Move APR constant to a separate constants file
