+++ src/composables/useNumbers.ts
@@ -20,10 +20,6 @@ enum PresetFormats {
   percent_lg = '0%'
 }
 
-// Do not display APR values greater than this amount; they are likely to be nonsensical
-// These can arise from pools with extremely low balances (e.g., completed LBPs)
-export const APR_THRESHOLD = 10000;
-
 export type Preset = keyof typeof PresetFormats;
 
 export function fNum(
